Split program dispatch runner for #3418
