+++ changelog.d/unreleased/3418.internal.md
@@ -1,0 +1,16 @@
+---
+category: internal
+issues:
+  - 3418
+affected:
+  - src/CodeIndex/Cli/ProgramRunner.cs
+  - src/CodeIndex/Cli/ProgramRunner.Dispatch.cs
+---
+
+## English
+
+- **Split top-level command dispatch into a focused partial (#3418)** — `ProgramRunner` now keeps immediate command handling, query dispatch, and non-query dispatch in a dedicated partial without changing CLI behavior.
+
+## 日本語
+
+- **top-level command dispatch を focused partial に分割しました (#3418)** — `ProgramRunner` の immediate command handling、query dispatch、non-query dispatch を専用 partial に移し、CLI 挙動は変更していません。
